feat: implement startServer function and update README with usage example
