+++ packages/agents-hosting-express/README.md
@@ -2,3 +2,17 @@
 
 ## Overview
 
+Provides integration to host the agent in Express using `startServer`
+
+## Usage
+
+ ```ts
+ import { AgentApplication, TurnState } from '@microsoft/agents-hosting';
+ import { startServer } from '@microsoft/agents-hosting-express';
+ 
+ const app = new AgentApplication<TurnState>();
+ app.message('hello', async (context, state) => {
+   await context.sendActivity('Hello, world!');
+ });
+ startServer(app);
+  ```
