+++ packages/agents-hosting-express/src/index.ts
@@ -1,1 +1,1 @@
-export * from './expressHost'
+export * from './startServer'
