+++ typedoc.json
@@ -4,6 +4,7 @@
       "./packages/agents-activity/index.ts",
       "./packages/agents-copilotstudio-client/index.ts",
       "./packages/agents-hosting/index.ts",
+      "./packages/agents-hosting-express/index.ts",
       "./packages/agents-hosting-dialogs/index.ts",
       "./packages/agents-hosting-storage-blob/index.ts",
       "./packages/agents-hosting-storage-cosmos/index.ts",
